pyproject.toml
[tool.poetry]
name = "src"
version = "1.0"
description = "Api for manor's owners."
authors = ["eugene <eugenestudio@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
attrs = "23.1.0"
certifi = "2023.5.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "41.0.1"
defusedxml = "0.7.1"
django = "3.2"
django-cors-headers = "3.14.0"
django-debug-toolbar = "3.4.0"
django-filter = "23.1"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
djoser = "2.2.0"
docutils = "0.19"
drf-spectacular = "0.26.4"
drf-yasg = "1.21.5"
idna = "3.4"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.2"
jsonschema = "4.17.3"
markupsafe = "2.1.2"
oauthlib = "3.2.2"
packaging = "23.1"
pillow = "9.5.0"
pycparser = "2.21"
pyjwt = "2.7.0"
pymysql = "1.0.3"
pyrsistent = "0.19.3"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3"
pyyaml = "6.0"
requests = "2.28.2"
requests-oauthlib = "1.3.1"
ruamel-yaml = "0.17.21"
social-auth-app-django = "5.2.0"
social-auth-core = "4.4.2"
sorl-thumbnail = "12.9.0"
sqlparse = "0.4.4"
tzdata = "2023.3"
uritemplate = "4.1.1"
urllib3 = "1.26.15"
whitenoise = "6.4.0"
django-phonenumber-field = "7.1.0"
phonenumbers = "8.13.18"
drf-spectacular-sidecar = "2023.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
